+++ src/framework/effects/Reverse.java
@@ -2,8 +2,8 @@ package framework.effects;
 
 public class Reverse implements EffectsProcessor {
 	
-	public double[] process(double[] buffer) {
-		double[] processedBuffer = new double[buffer.length];
+	public float[] process(float[] buffer) {
+		float[] processedBuffer = new float[buffer.length];
 		for (int i = 1; i <= buffer.length; i++) {
 			processedBuffer[i - 1] = buffer[buffer.length - i];
 		}
